Fill Matrix-Of-Clusters

This function fills in a matrix of clusters that contains NAs, by estimating the missing cluster labels based on the available ones.

fillMOC(clLabels, computeAccuracy = FALSE, verbose = FALSE)

Arguments

clLabels
N X M matrix containing cluster labels. Element (n,m) contains the cluster label for element data point n in cluster m.

computeAccuracy
Boolean. If TRUE, for each missing element, the performance of the predictive model used to estimate the corresponding missing label is computer.

verbose
Boolean. If TRUE, for each NA, the size of the matrix used to estimate its values is printed to screen

Value

The output is the same matrix of clusters, where NAs have been replaced by their estimates, where possible. If `computeAccuracy` is TRUE, then also an object called `accuracy` is returned, where each

References

The Cancer Genome Atlas, 2012. Comprehensive molecular portraits of human breast tumours. Nature, 487(7407), pp.61–70.

Examples

## Load data
data <- list()
data[[1]] <- as.matrix(read.csv(system.file("extdata", "dataset1.csv",
package = "coca"), row.names = 1))
data[[2]] <- as.matrix(read.csv(system.file("extdata", "dataset2.csv",
package = "coca"), row.names = 1))
data[[3]] <- as.matrix(read.csv(system.file("extdata", "dataset3.csv",
package = "coca"), row.names = 1))

## Build matrix of clusters
outputBuildMOC <- buildMOC(data, M = 3, K = 6)

## Extract matrix of clusters
clLabels <- outputBuildMOC$clLabels

## Impute missing values
outputFillMOC <- fillMOC(clLabels)

## Replace matrix of cluster labels with new (full) one
clLabels <- outputFillMOC$fullClLabels